Added files for client-side logic.
Need to refactor structure to ensure that only one Controller and
PeerConnectionLayer are spawned. A Controller should be equipped with
multiple CRDT data structures (one per Atom 'TextBuffer').

Currently, on Controller and one PeerConnectionLayer are spawned per
Atom 'TextEditor' instance. This is undesirable since we spawn a
connection to the server (and a different P2P connection) for each
'TextEditor' instance.