WIP: Last commit before beginning workflow refactor. TODO:
1. When `HostPortalBinding` is created, a new `TextEditor` and new
`TextBuffer` should be created. This `TextBuffer` will be the single
`TextBuffer` connected to the `HostPortalBinding` (i.e. the only `TextBuffer`
involved in interactions with portal guests). We'll call this
`TextBuffer` the "portal-host TextBuffer".

2. When a `HostPortalBinding` transmits the `JOIN_REQUEST_ACCEPTED`
message to a remote peer, the `HostPortalBinding` should also send the
ID of the `TextBufferProxy` that holds the portal-host `TextBuffer`.

3. When the remote peer delivers the `JOIN_REQUEST_ACCEPTED` message
from the portal host, it should convert the ID (in a clear and
systematic way) of the portal-host `TextBufferProxy` to one for a
portal-guest. This ID will then be used to set the path for the
portal-guest `TextBuffer` and create the portal-guest `TextBufferProxy`
in which the portal-guest `TextBuffer` will be contained.

4. The `HostPortalBinding` should do the necessary `TextBufferProxy` ID
conversions when delivering a text-insertion/text-deletion message from
a portal guest.